restructured project, added relative imports, fixed namespace issues
