+++ src/lib/utils/SynapseConstants.js
