+++ src/test/containers/App.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './App';
+import App from 'containers/App';
 
 it('renders without crashing', () => {
   const div = document.createElement('div');
